pyproject.toml
[project]
name = "bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.24.0",
    "alembic>=1.18.1",
    "anthropic>=0.76.0",
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "greenlet>=3.3.1",
    "loguru>=0.7.3",
    "openai>=2.15.0",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
